add reindexer for sms pillow
