+++ corehq/apps/hqcase/management/commands/ptop_preindex.py
@@ -33,7 +33,8 @@ def get_reindex_commands(alias_name):
         'hqgroups': [('ptop_reindexer_v2', {'index': 'group'})],
         'report_xforms': ['ptop_fast_reindex_reportxforms'],
         'report_cases': ['ptop_fast_reindex_reportcases'],
-        'case_search': [('ptop_reindexer_v2', {'index': 'case-search'})]
+        'case_search': [('ptop_reindexer_v2', {'index': 'case-search'})],
+        'smslogs': [('ptop_reindexer_v2', {'index': 'sms'})],
     }
     return pillow_command_map.get(alias_name, [])
