+++ corehq/apps/hqcase/management/commands/ptop_reindexer_v2.py
@@ -6,6 +6,7 @@ from corehq.pillows.domain import get_domain_reindexer
 from corehq.pillows.group import get_group_reindexer
 from corehq.pillows.groups_to_user import get_groups_to_user_reindexer
 from corehq.pillows.ledger import get_ledger_v2_reindexer, get_ledger_v1_reindexer
+from corehq.pillows.sms import get_sms_reindexer
 from corehq.pillows.user import get_user_reindexer
 from corehq.pillows.xform import get_couch_form_reindexer, get_sql_form_reindexer
 
@@ -42,6 +43,7 @@ class Command(BaseCommand):
             'case-search': get_case_search_reindexer,
             'ledger-v2': get_ledger_v2_reindexer,
             'ledger-v1': get_ledger_v1_reindexer,
+            'sms': get_sms_reindexer,
         }
         if index not in reindex_fns:
             raise CommandError('Supported indices to reindex are: {}'.format(','.join(reindex_fns.keys())))
