+++ corehq/form_processor/change_providers.py
@@ -70,30 +70,6 @@ def _ledger_v2_to_change(ledger_value):
     )
 
 
-class DjangoModelChangeProvider(ChangeProvider):
-
-    def __init__(self, model_class, model_to_change_fn, chunk_size=500):
-        self.model_class = model_class
-        self.model_to_change_fn = model_to_change_fn
-        self.chunk_size = chunk_size
-
-    def iter_all_changes(self, start_from=None):
-        from django.core.paginator import Paginator
-        from django.core.paginator import EmptyPage
-
-        model_list = self.model_class.objects.all()
-        paginator = Paginator(model_list, self.chunk_size)
-
-        page = 0
-        while True:
-            page += 1
-            try:
-                for model in paginator.page(page):
-                    yield self.model_to_change_fn(model)
-            except EmptyPage:
-                return
-
-
 def _ledger_v1_to_change(stock_state):
     return Change(
         id=stock_state.pk,
