+++ corehq/pillows/sms.py
@@ -1,17 +1,20 @@
 from corehq.apps.change_feed import topics
 from corehq.apps.change_feed.consumer.feed import KafkaChangeFeed
+from corehq.apps.sms.change_publishers import change_meta_from_sms
 from corehq.elastic import get_es_new
 from corehq.pillows.mappings.sms_mapping import SMS_INDEX_INFO
 from pillowtop.checkpoints.manager import PillowCheckpoint, PillowCheckpointEventHandler
+from pillowtop.feed.interface import Change
 from pillowtop.pillow.interface import ConstructedPillow
 from pillowtop.processors.elastic import ElasticProcessor
-
+from pillowtop.reindexer.change_providers.django_model import DjangoModelChangeProvider
+from pillowtop.reindexer.reindexer import ElasticPillowReindexer
 
 SMS_PILLOW_CHECKPOINT_ID = 'sql-sms-to-es'
 SMS_PILLOW_KAFKA_CONSUMER_GROUP_ID = 'sql-sms-to-es'
 
 
-def get_sql_sms_pillow(pillow_id):
+def get_sql_sms_pillow(pillow_id='sql-sms-pillow'):
     checkpoint = PillowCheckpoint(SMS_PILLOW_CHECKPOINT_ID)
     processor = ElasticProcessor(
         elasticsearch=get_es_new(),
@@ -26,4 +29,25 @@ def get_sql_sms_pillow(pillow_id):
         change_processed_event_handler=PillowCheckpointEventHandler(
             checkpoint=checkpoint, checkpoint_frequency=100,
         ),
+    )
+
+
+def get_sms_reindexer():
+    from corehq.apps.sms.models import SMS
+    return ElasticPillowReindexer(
+        pillow=get_sql_sms_pillow(),
+        change_provider=DjangoModelChangeProvider(SMS, _sql_sms_to_change),
+        elasticsearch=get_es_new(),
+        index_info=SMS_INDEX_INFO,
+    )
+
+
+def _sql_sms_to_change(sms):
+    return Change(
+        id=sms.couch_id,
+        sequence_id=None,
+        document=sms.to_json(),
+        deleted=False,
+        metadata=change_meta_from_sms(sms),
+        document_store=None,
     )
